Add Private Chat models
Private_chat model, represents individual chat sessions or chat channels between two profiles. It has a one-to-many relationship with the Message model, which represents the messages sent in the chat session. The Private_chat model also has a many-to-many relationship with the Profile model, which represents the profiles that are part of the chat session.